Remove android:allowBackup="true"
Allow apps using the library to select their own backup option without needing to override it. Libraries should't use allowBackup in theeir manifest. This provokes a build error in recent versions of the Gradle plugin when the application sets this value to false.

Temporary workaround is to add `tools:replace="android:allowBackup"` to the *application* tag in the app's manifest.